Models necessary for the collapsed-tunnel map.
Models made for the test map, "collapsed-tunnel".

Textures were sourced from CC0textures for the models "Collapsed-tunnel" and "Grass1". The road texture is a mixture of two of them.
"Waypoint-pole" is a very old and low poly model i made years ago, but was used on the collapsed tunnel map.